Test submimssion commit: id=24798380, title=하노이 탑 이동 순서

+++ baekjoon/11729.cpp
@@ -29,6 +29,6 @@ int main() {
 	int n;
 	cin >> n;
 
-	cout << pow(2, n) - 1 << "\n";
+	cout << (int)pow(2, n) - 1 << "\n";
 	hanoi(n, 1, 3, 2);
 }
